refactor: move input to separate component
